refactor: improve code quality
